requirements/prod.txt
# Include base requirements
-r base.txt

# Production ASGI Server
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# Error Monitoring & Observability
sentry-sdk[sqlalchemy,fastapi]==1.38.0

# Metrics & Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# APM Monitoring
elastic-apm==6.20.0

# Production Database Connection Pooling
psycopg2-binary==2.9.9
pgbouncer==1.21.0

# Production Caching
redis[hiredis]==5.0.1

# Production Logging
structlog==23.2.0
python-json-logger==2.0.7

# Security Enhancements
cryptography==41.0.8
argon2-cffi==23.1.0

# Rate Limiting
slowapi==0.1.9

# CORS & Security Headers
python-multipart==0.0.6

# Health Checks
healthcheck==1.3.3

# Process Management
supervisor==4.2.5

# Memory Management
psutil==5.9.6
memory-profiler==0.61.0

# Production Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File Storage (Production)
boto3==1.34.0
botocore==1.34.0

# Email Services (Production)
sendgrid==6.10.0
mailgun2==1.0.0

# SMS Services (Production)
twilio==8.11.0

# CDN & Static Files
whitenoise==6.6.0

# Database Optimization
sqlalchemy-utils==0.41.1

# Production Task Queue
celery[redis]==5.3.4
flower==2.0.1

# Backup & Recovery
pg-dump==1.0.0

# SSL/TLS Support
certifi==2023.11.17

# Production Utilities
click==8.1.7
python-dateutil==2.8.2

# Network Security
requests[security]==2.31.0
urllib3[secure]==2.1.0

# Production JSON Processing
orjson==3.9.10

# Timezone Support
pytz==2023.3
tzdata==2023.3

# Production File Processing
python-magic==0.4.27

# Image Optimization
pillow-simd==10.0.1.post1

# Production Analytics
segment-analytics-python==2.2.3

# Circuit Breaker Pattern
pybreaker==1.0.1

# Retry Logic
tenacity==8.2.3

# Production Validation
cerberus==1.3.5

# Environment Detection
python-decouple==3.8

# Production Secrets Management
keyring==24.3.0
keyrings.cryptfile==1.3.9

# Background Task Monitoring
celery-redbeat==2.2.0

# Production Queue Management
kombu==5.3.4

# SSL Certificate Management
acme==2.7.4

# Production Database Migrations
alembic==1.13.0

# Production Backup
duplicity==2.1.4

# System Monitoring
docker==6.1.3

# Production Deployment
fabric==3.2.2

# Container Health
docker-healthcheck==1.0.0

# Production Cache
redis-py-cluster==2.1.3

# Load Balancing
haproxy-stats==2.4.22

# Production Analytics
mixpanel==4.10.0

# Feature Flags
flagsmith==3.7.0

# A/B Testing
optimizely-sdk==4.1.1

# Production Notifications
slack-sdk==3.25.0
discord-webhook==1.3.1

# Production File Storage
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Production Search
elasticsearch==8.11.0

# Production Message Queue
pika==1.3.2
confluent-kafka==2.3.0

# Production Cron Jobs
django-crontab==0.7.1
crontab==1.0.1

# Production Process Monitoring
supervisor==4.2.5
circus==0.18.0

# Production Memory Cache
pylibmc==1.6.3
python-memcached==1.59

# Production WebSocket Support
websockets==12.0

# Production API Rate Limiting
limits==3.6.0

# Production Security
authlib==1.2.1
PyJWT==2.8.0

# Production Data Validation
marshmallow==3.20.2